Divy up middleware into seperate files
