+++ lib/middleware/bowerComponents.js
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const bowerComponents = express.static('bower_components');
+
+export default bowerComponents;
