+++ lib/middleware/notFoundError.js
@@ -1,0 +1,5 @@
+export default function notFoundError(req, res, next) {
+  const err = new Error('Not Found');
+  err.status = 404;
+  next(err);
+}
