+++ lib/middleware/serverError.js
@@ -1,0 +1,4 @@
+export default function serverError(err, req, res) {
+  res.status(err.status || 500);
+  res.send(err);
+}
