+++ lib/server.js
@@ -2,13 +2,10 @@ import express from 'express';
 import figshareRoute from './routes/figshare';
 import dataOneRoute from './routes/dataone';
 import dataciteRoute from './routes/datacite';
-
-import {
-  notFoundError,
-  serverError,
-  staticAssets,
-  bowerComponents,
-} from './Middleware';
+import notFoundError from './middleware/notFoundError';
+import serverError from './middleware/serverError';
+import staticAssets from './middleware/staticAssets';
+import bowerComponents from './middleware/bowerComponents';
 
 const app = express();
